Adding some genericity and add benchmark class
A new class AbstractCentrality can be used to evaluate graphs with
different types of graphs.

GraphListType is used to change the list type for graphs (ArrayList<>
here)

A new class "Benchmark" has been created to perform benchmark, such as
get elapsed time between two "snapshots" and get iteration count.
The tutorial will be updated to explain how to use this